refactor(Card): remove card component
